requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
python-dotenv==1.0.0
reportlab==4.0.8
pydantic==2.4.2
numpy==1.24.3
scikit-learn==1.3.0
huggingface-hub==0.16.4